deletion queue: refactor into frontend/backend modules
